+++ src/modules/featureInfo/components/featureInfoPanel/featureInfoPanel.css
